Split models.py into separate modules for lock, code, log
