+++ tests/test_code.py
@@ -1,0 +1,125 @@
+from copy import deepcopy
+from datetime import datetime
+from unittest import mock
+
+import pyschlage
+from pyschlage.code import AccessCode, DaysOfWeek, RecurringSchedule, TemporarySchedule
+
+from .fixtures import ACCESS_CODE_JSON
+
+
+class TestAccessCode:
+    def test_to_from_json(self):
+        auth = mock.Mock()
+        device_id = "__device_uuid__"
+        access_code_id = "__access_code_uuid__"
+        code = AccessCode(
+            _auth=auth,
+            name="Friendly name",
+            code="0123",
+            schedule=None,
+            device_id=device_id,
+            access_code_id=access_code_id,
+        )
+        assert AccessCode.from_json(auth, ACCESS_CODE_JSON, device_id) == code
+
+        want_json = deepcopy(ACCESS_CODE_JSON)
+        # We don't send back the id because it's not mutable.
+        del want_json["accesscodeId"]
+        assert code.to_json() == want_json
+
+    def test_to_from_json_recurring_schedule(self):
+        auth = mock.Mock()
+        device_id = "__device_uuid__"
+        access_code_id = "__access_code_uuid__"
+        sched = RecurringSchedule(days_of_week=DaysOfWeek(mon=False))
+        json = deepcopy(ACCESS_CODE_JSON)
+        json["schedule1"] = sched.to_json()
+        code = AccessCode(
+            _auth=auth,
+            name="Friendly name",
+            code="0123",
+            schedule=sched,
+            device_id=device_id,
+            access_code_id=access_code_id,
+        )
+        assert AccessCode.from_json(auth, json, device_id) == code
+
+        # We don't send back the id because it's not mutable.
+        del json["accesscodeId"]
+        assert code.to_json() == json
+
+    def test_to_from_json_temporary_schedule(self):
+        auth = mock.Mock()
+        device_id = "__device_uuid__"
+        access_code_id = "__access_code_uuid__"
+        sched = TemporarySchedule(
+            start=datetime(2022, 12, 25, 8, 30, 0),
+            end=datetime(2022, 12, 25, 9, 0, 0),
+        )
+        json = deepcopy(ACCESS_CODE_JSON)
+        json["activationSecs"] = 1671957000
+        json["expirationSecs"] = 1671958800
+        code = AccessCode(
+            _auth=auth,
+            name="Friendly name",
+            code="0123",
+            schedule=sched,
+            device_id=device_id,
+            access_code_id=access_code_id,
+        )
+        assert AccessCode.from_json(auth, json, device_id) == code
+
+        # We don't send back the id because it's not mutable.
+        del json["accesscodeId"]
+        assert code.to_json() == json
+
+    def test_refresh(self):
+        auth = mock.Mock()
+        code = AccessCode.from_json(auth, ACCESS_CODE_JSON, "__device_uuid__")
+        new_json = deepcopy(ACCESS_CODE_JSON)
+        new_json["accessCode"] = 1122
+
+        auth.request.return_value = mock.Mock(json=mock.Mock(return_value=new_json))
+        code.refresh()
+
+        auth.request.assert_called_once_with(
+            "get", "devices/__device_uuid__/storage/accesscode/__access_code_uuid__"
+        )
+        assert code.code == "1122"
+
+    def test_save(self):
+        auth = mock.Mock()
+        code = AccessCode.from_json(auth, ACCESS_CODE_JSON, "__device_uuid__")
+        code.code = 1122
+        old_json = code.to_json()
+
+        new_json = deepcopy(ACCESS_CODE_JSON)
+        new_json["accessCode"] = 1122
+        # Simulate another change that happened out of band.
+        new_json["friendlyName"] = "New name"
+
+        auth.request.return_value = mock.Mock(json=mock.Mock(return_value=new_json))
+        code.save()
+
+        auth.request.assert_called_once_with(
+            "put",
+            "devices/__device_uuid__/storage/accesscode/__access_code_uuid__",
+            json=old_json,
+        )
+        assert code.code == "1122"
+        # Ensure the name was updated.
+        assert code.name == "New name"
+
+    def test_delete(self):
+        auth = mock.Mock()
+        code = AccessCode.from_json(auth, ACCESS_CODE_JSON, "__device_uuid__")
+        auth.request.return_value = mock.Mock()
+        code.delete()
+        auth.request.assert_called_once_with(
+            "delete", "devices/__device_uuid__/storage/accesscode/__access_code_uuid__"
+        )
+        assert code._auth is None
+        assert code.access_code_id is None
+        assert code.device_id is None
+        assert code.disabled == True
